@@ -2,33 +2,61 @@
   "name": "tremors-node",
   "version": "1.0.0",
   "description": "",
-  "main": "src/app.js",
+  "main": "src/server/app.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest",
+    "dev-server": "nodemon --exec ts-node src/server/app.ts --ignore dist/",
+    "dev-bundle": "webpack -wd"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@babel/core": "^7.7.7",
+    "@babel/node": "^7.7.7",
+    "@babel/preset-env": "^7.7.7",
+    "@babel/preset-react": "^7.7.4",
+    "@react-google-maps/api": "^1.8.1",
+    "axios": "^0.19.0",
+    "babel-loader": "^8.0.6",
     "body-parser": "^1.19.0",
     "compression": "^1.7.4",
-    "config": "^3.2.3",
+    "config": "^3.2.4",
     "express": "^4.17.1",
-    "express-validator": "^6.2.0",
-    "faker": "^4.1.0",
-    "ioredis": "^4.14.1",
+    "express-validator": "^6.3.1",
+    "google-map-react": "^1.1.5",
     "lusca": "^1.6.1",
     "morgan": "^1.9.1",
-    "pug": "^2.0.4"
+    "pug": "^2.0.4",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "webpack": "^4.41.5",
+    "webpack-cli": "^3.3.10"
   },
   "devDependencies": {
+    "@babel/preset-typescript": "^7.7.7",
+    "@types/body-parser": "^1.17.1",
+    "@types/compression": "^1.0.1",
+    "@types/express": "^4.17.2",
+    "@types/lusca": "^1.6.1",
+    "@types/morgan": "^1.7.37",
+    "@types/node": "^12.12.22",
+    "babel-eslint": "^10.0.3",
+    "babel-jest": "^24.9.0",
     "chai": "^4.2.0",
-    "eslint": "^6.5.1",
+    "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.0.0",
     "eslint-plugin-chai-friendly": "^0.4.1",
-    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-import": "^2.19.1",
+    "eslint-plugin-react": "^7.17.0",
+    "eslint-plugin-react-hooks": "^2.3.0",
+    "jest": "^24.9.0",
     "mocha": "^6.2.1",
+    "nodemon": "^1.19.4",
     "nyc": "^14.1.1",
+    "react-test-renderer": "^16.12.0",
     "sinon": "^7.5.0",
-    "supertest": "^4.0.2"
+    "supertest": "^4.0.2",
+    "ts-node": "^8.5.4",
+    "typescript": "^3.7.4"
   }
 }
